Put coverage report production code in a script file
Recently, GDB has started printing one ">" prompt per line in the
"python" command. For instance, when the input is:

    python
    pass
    pass
    pass
    end

GDB prints:

    >>>(gdb)
    [plus a trailing space, i.e. the main prompt]

Since this breaks the current regexp that matches GDB prompts ("^(gdb)
", so the pexpect machinery hangs), switch to using a script file
instead to workaround this problem.

TN: U622-042
Change-Id: I0e6d6cc23a6b8ed350c1bc695bbd79913a8adecf