Simplify registration and make printers Ada-specific
This patch simplifies pretty-printer registration.  In particular,
they are now simply registered globally.

In conjunction with this, the printers are now changed to only
activate when the current language is Ada.  This requires a small gdb
patch to expose this information; there is a patch pending for this.

I examined a few different approaches here, becase there are two
different problems to avoid.  One is that, with the new DWARF reader,
it's best to defer symbol lookups as long as possible -- a symbol
lookup requires waiting for the DWARF scan to finish, which introduces
a user-visible pause.  The second problem is that, because many Ada
programs are statically linked, the usual approach of a .py file
associated with a shared library would not work.  (We could use a
.debug_gdb_scripts section, but this has other issues.)

I settled on this approach as it seems to be reasonable, and provides
what users probably want anyway.

Change-Id: I40a598272143fd61ec3be685906f3b4812781907
TN: V518-040