Simplify string printing
Right now, gnat-gdb-scripts implements its own string formatting code,
I think so that it can also show the type and address in the output.
However, this is a bit verbose, and the experience with the C++
printers is that just showing the string data seems fine for users.

This patch simplifies the string printing to use lazy_string and use
the "string" display hint.  If we do want to restore the type and/or
address, it seems preferable to add this to gdb proper.

Change-Id: Id06781a3ae52bddc88a4c75ed8fa42de30ba15d8
no-tn-check